+++ src/ast.rs
@@ -1,0 +1,11 @@
+mod exp;
+mod statement;
+
+pub use exp::*;
+pub use statement::*;
+
+pub enum AST {
+	StmtNode(Statement, Box<AST>),
+	ExpNode(Exp, Box<AST>),
+	Leaf,
+}
