+++ src/lib.rs
@@ -1,3 +1,4 @@
 mod parser;
 mod tokenizer;
+mod ast;
 mod types;
