--- src/parser/ast.rs
@@ -1,8 +1,0 @@
-use super::exp::Exp;
-use super::statement::Statement;
-
-pub enum AST {
-	NodeStmt(Statement, Box<AST>),
-	NodeExp(Exp, Box<AST>),
-	Leaf,
-}
