+++ src/tokenizer.rs
@@ -1,2 +1,4 @@
-pub mod token;
+mod token;
 pub mod tokenizer;
+
+pub use token::Token;
